requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.5
django-filter==22.1
django-heroku==0.3.1
django-soft-delete==0.9.21
gunicorn==20.1.0
idna==3.3
msal==1.18.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
requests==2.28.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
